chore: code split for approval page crud modal (#39615)
## Description
This PR splits the modals and the context menu hooks to allow for
enterprise only features to show on the canvas. No change in the code,
just re-arrangement of files. Current UTs and tests should suffice


Fixes #39640 

## Automation

/test all

### :mag: Cypress test results
